refactor(phase-3A): convert cli.py to package, extract migrate + lineage modules [no behavior change]
First chunk of the cli.py split per ari-core/REFACTORING.md §3.

Structural change:
- ``ari/cli.py`` → ``ari/cli/__init__.py`` (git mv) so the planned
  ``ari/cli/{lineage,migrate,run,projects,commands,bfts_loop}.py``
  package can grow alongside it.
- New ``ari/cli/__main__.py`` keeps ``python -m ari.cli`` working
  after the package conversion.

Sub-modules extracted in this PR:
- ``ari/cli/migrate.py`` — ``cmd_migrate_node_reports`` + the
  ``migrate_app`` Typer sub-app.  ``__init__.py`` reduces to
  ``from ari.cli.migrate import migrate_app, cmd_migrate_node_reports``
  + ``app.add_typer(migrate_app, name="migrate")``.
- ``ari/cli/lineage.py`` — ``_LINEAGE_LOG`` /
  ``_load_lineage_decision_config`` / ``_mark_parent_terminated`` /
  ``_execute_lineage_decision`` / ``_build_idea_ctx_for_expand``;
  ``__init__.py`` re-imports the same names so ``_run_loop`` keeps
  calling them by their original names.

Bundled-config path fix:
- After the package conversion ``__file__`` gains an extra path
  segment (``ari/cli/__init__.py`` instead of ``ari/cli.py``).  The
  five ``Path(__file__).parent.parent / "config"`` references in
  ``cli/__init__.py`` now go up one more level
  (``parent.parent.parent``) so the bundled
  ``ari-core/config/{workflow,profiles/*}.yaml`` files keep
  resolving — Phase 2 ``ari.config.finder.package_config_root`` will
  collapse these into one helper in a follow-up.

Test fixture updates:
- ``test_config.py:test_profile_file_is_findable`` now walks up an
  extra ``__file__`` parent to land on ``ari-core/`` (Phase 3A
  package layout).
- 8 test sites that read ``ari/cli.py`` as text (``test_idea_integration``,
  ``test_pipeline_e2e``, ``test_wizard``, ``test_server``) read
  ``ari/cli/__init__.py`` instead so the in-flight code-presence
  checks still pass.

Verification:
- All 13 subcommand ``--help`` outputs (root, run, resume, paper,
  viz, status, projects, show, delete, settings, skills-list, clone,
  migrate) match the pre-Phase-3A baseline byte-for-byte.
- ``pytest ari-core/tests/ -q`` (excluding live integrations):
  1944 passed, 15 skipped.

Phase 3A continues in follow-up commits: the larger run/resume/paper
/projects/commands/bfts_loop modules still live in ``__init__.py``.
